sdk/txbuild: move integration test helpers into another file
